Refactor FE
